mcp_client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.12.5",
    "google-genai>=1.46.0",
    "langgraph>=1.0.2",
    "rich>=14.2.0",
]
